feat: 작성위치 잘못으로 인한 test에서 main으로 이동 (#2)
